python-producer/requirements.txt
grpcio
grpcio-tools
pandas
sdv
protobuf
mesa